removed subfolder structure of qpsolver
