+++ src/qpsolver/devexpricing.hpp
@@ -1,16 +1,18 @@
-#ifndef __SRC_LIB_PRICING_STEEPESTEDGEPRICING_HPP__
-#define __SRC_LIB_PRICING_STEEPESTEDGEPRICING_HPP__
+#ifndef __SRC_LIB_PRICING_DEVEXPRICING_HPP__
+#define __SRC_LIB_PRICING_DEVEXPRICING_HPP__
 
-#include "../basis.hpp"
-#include "../runtime.hpp"
+#include "basis.hpp"
+#include "reducedcosts.hpp"
+#include "runtime.hpp"
 #include "pricing.hpp"
 
-//
+// 42726, 78965776.391299, 559, 104.321553, 0.000669, 7937
 
-class SteepestEdgePricing : public Pricing {
+class DevexPricing : public Pricing {
  private:
   Runtime& runtime;
   Basis& basis;
+  ReducedCosts& redcosts;
 
   std::vector<double> weights;
 
@@ -19,7 +21,7 @@ class SteepestEdgePricing : public Pricing {
     auto constraintindexinbasisfactor = basis.getindexinfactor();
 
     HighsInt minidx = -1;
-    double maxval = 0.0;
+    double maxabslambda = 0.0;
     for (HighsInt i = 0; i < activeconstraintidx.size(); i++) {
       HighsInt indexinbasis =
           constraintindexinbasisfactor[activeconstraintidx[i]];
@@ -30,18 +32,18 @@ class SteepestEdgePricing : public Pricing {
 
       double val = lambda.value[indexinbasis] * lambda.value[indexinbasis] /
                    weights[indexinbasis];
-      if (val > maxval && fabs(lambda.value[indexinbasis]) >
-                              runtime.settings.lambda_zero_threshold) {
+      if (val > maxabslambda && fabs(lambda.value[indexinbasis]) >
+                                    runtime.settings.lambda_zero_threshold) {
         if (basis.getstatus(activeconstraintidx[i]) ==
                 BasisStatus::ActiveAtLower &&
             -lambda.value[indexinbasis] > 0) {
           minidx = activeconstraintidx[i];
-          maxval = val;
+          maxabslambda = val;
         } else if (basis.getstatus(activeconstraintidx[i]) ==
                        BasisStatus::ActiveAtUpper &&
                    lambda.value[indexinbasis] > 0) {
           minidx = activeconstraintidx[i];
-          maxval = val;
+          maxabslambda = val;
         } else {
           // TODO
         }
@@ -52,13 +54,21 @@ class SteepestEdgePricing : public Pricing {
   }
 
  public:
-  SteepestEdgePricing(Runtime& rt, Basis& bas)
+  DevexPricing(Runtime& rt, Basis& bas, ReducedCosts& rc)
       : runtime(rt),
         basis(bas),
+        redcosts(rc),
         weights(std::vector<double>(rt.instance.num_var, 1.0)){};
 
+  // B lambda = g
+  // lambda = inv(B)g
+  // lambda = Z'g == reduced gradient ??
+  // no: lambda = Y'g !!
+  // dual values updated as:
+  // c_N^T  += alpha_D * a_p^T (pivotal row)
+  // alpha_D = -c_q / a_pq
   HighsInt price(const Vector& x, const Vector& gradient) {
-    Vector lambda = basis.ftran(gradient);
+    Vector& lambda = redcosts.getReducedCosts();
     HighsInt minidx = chooseconstrainttodrop(lambda);
     return minidx;
   }
@@ -66,19 +76,17 @@ class SteepestEdgePricing : public Pricing {
   void update_weights(const Vector& aq, const Vector& ep, HighsInt p,
                       HighsInt q) {
     HighsInt rowindex_p = basis.getindexinfactor()[p];
-
-    Vector v = basis.btran(aq);
-
     double weight_p = weights[rowindex_p];
     for (HighsInt i = 0; i < runtime.instance.num_var; i++) {
       if (i == rowindex_p) {
         weights[i] = weight_p / (aq.value[rowindex_p] * aq.value[rowindex_p]);
       } else {
-        weights[i] = weights[i] -
-                     2 * (aq.value[i] / aq.value[rowindex_p]) * (v.value[i]) +
-                     (aq.value[i] * aq.value[i]) /
-                         (aq.value[rowindex_p] * aq.value[rowindex_p]) *
-                         weight_p;
+        weights[i] += (aq.value[i] * aq.value[i]) /
+                      (aq.value[rowindex_p] * aq.value[rowindex_p]) * weight_p *
+                      weight_p;
+      }
+      if (weights[i] > 10E6) {
+        weights[i] = 1.0;
       }
     }
   }
