+++ src/qpsolver/pricing.hpp
@@ -1,8 +1,8 @@
 #ifndef __SRC_LIB_PRICING_HPP__
 #define __SRC_LIB_PRICING_HPP__
 
-#include "../matrix.hpp"
-#include "../vector.hpp"
+#include "matrix.hpp"
+#include "vector.hpp"
 
 class Pricing {
  public:
