+++ src/qpsolver/ratiotest.cpp
