+++ src/qpsolver/solver.cpp
@@ -5,17 +5,17 @@
 
 #include "Highs.h"
 #include "basis.hpp"
-#include "feasibility/feasibility_highs.hpp"
+#include "feasibility_highs.hpp"
 #include "gradient.hpp"
 #include "instance.hpp"
 #include "lp_data/HighsAnalysis.h"
 #include "nullspace.hpp"
-#include "pricing/dantzigpricing.hpp"
-#include "pricing/devexharrispricing.hpp"
-#include "pricing/devexpricing.hpp"
-#include "pricing/steepestedgepricing.hpp"
-#include "ratiotest/ratiotest.hpp"
-#include "redhes/factor.hpp"
+#include "dantzigpricing.hpp"
+#include "devexharrispricing.hpp"
+#include "devexpricing.hpp"
+#include "steepestedgepricing.hpp"
+#include "ratiotest.hpp"
+#include "factor.hpp"
 #include "reducedcosts.hpp"
 #include "reducedgradient.hpp"
 #include "snippets.hpp"
